+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Allow BitBar Selenium server URL to be configured [524](https://github.com/bugsnag/maze-runner/pull/524)
 - Improve JSON schema validation failure messages [516](https://github.com/bugsnag/maze-runner/pull/516)
 
+## Fixes
+
+- Refactor schema validation and add it to span expectation step [520](https://github.com/bugsnag/maze-runner/pull/520)
+
 ## Removals
 
 - Remove absolute browser versions no longer supported by BitBar [523](https://github.com/bugsnag/maze-runner/pull/523)